requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
composio-core>=0.4.0
python-dotenv>=1.0.0

